requirements.txt
easyocr
pillow
pyperclip
